Reworked compressed and protocol implementation,   including fixes for conc-31 and conc-34 - Added win64 fixes in protocol (changed ulong to size_t)
modified:
  include/my_sys.h
  include/mysql_com.h
  include/violite.h
  libmariadb/libmariadb.c
  libmariadb/my_compress.c
  libmariadb/my_stmt.c
  libmariadb/my_thr_init.c
  libmariadb/net.c
  libmariadb/violite.c
  unittest/libmariadb/basic-t.c
  unittest/libmariadb/charset.c
  unittest/libmariadb/thread.c
unknown:
  xx
  libmariadb/libmariadb.so.1
  mariadb_config/mariadb_config
  mariadb_config/mariadb_config.c
  unittest/libmariadb/basic-t
  unittest/libmariadb/charset
  unittest/libmariadb/connection
  unittest/libmariadb/cursor
  unittest/libmariadb/errors
  unittest/libmariadb/fetch
  unittest/libmariadb/logs
  unittest/libmariadb/misc
  unittest/libmariadb/ps
  unittest/libmariadb/ps_bugs
  unittest/libmariadb/ps_new
  unittest/libmariadb/result
  unittest/libmariadb/sp
  unittest/libmariadb/sqlite3
  unittest/libmariadb/ssl
  unittest/libmariadb/thread
  unittest/libmariadb/view

+++ include/my_sys.h
@@ -616,9 +616,9 @@ void load_defaults(const char *conf_file, const char **groups,
 		   int *argc, char ***argv);
 void free_defaults(char **argv);
 void print_defaults(const char *conf_file, const char **groups);
-my_bool my_compress(unsigned char *, ulong *, ulong *);
-my_bool my_uncompress(unsigned char *, ulong *, ulong *);
-unsigned char *my_compress_alloc(const unsigned char *packet, ulong *len, ulong *complen);
+my_bool my_compress(unsigned char *, size_t *, size_t *);
+my_bool my_uncompress(unsigned char *, size_t *, size_t *);
+unsigned char *my_compress_alloc(const unsigned char *packet, size_t *len, size_t *complen);
 ulong checksum(const unsigned char *mem, uint count);
 
 #if defined(_MSC_VER) && !defined(_WIN32)
