+++ libmariadb/my_compress.c
@@ -29,7 +29,7 @@
 ** *complen is 0 if the packet wasn't compressed
 */
 
-my_bool my_compress(unsigned char *packet, ulong *len, ulong *complen)
+my_bool my_compress(unsigned char *packet, size_t *len, size_t *complen)
 {
   if (*len < MIN_COMPRESS_LENGTH)
     *complen=0;
@@ -39,12 +39,13 @@ my_bool my_compress(unsigned char *packet, ulong *len, ulong *complen)
     if (!compbuf)
       return *complen ? 0 : 1;
     memcpy(packet,compbuf,*len);
-    my_free(compbuf,MYF(MY_WME));						  }
+    my_free(compbuf,MYF(MY_WME));
+  }
   return 0;
 }
 
 
-unsigned char *my_compress_alloc(const unsigned char *packet, ulong *len, ulong *complen)
+unsigned char *my_compress_alloc(const unsigned char *packet, size_t *len, size_t *complen)
 {
   unsigned char *compbuf;
   *complen =  *len * 120 / 100 + 12;
@@ -67,7 +68,7 @@ unsigned char *my_compress_alloc(const unsigned char *packet, ulong *len, ulong 
 }
 
 
-my_bool my_uncompress (unsigned char *packet, ulong *len, ulong *complen)
+my_bool my_uncompress (unsigned char *packet, size_t *len, size_t *complen)
 {
   if (*complen)					/* If compressed */
   {
